Unbreak the MacOS tests
Summary: The tests are currently broken by a new check added in D18200. in order for this check to succeed all the time we need the QT settings file to use an absolute path and there are some backports missing to achieve this, so we disable the check for now.

Test Plan:
  ninja check-bitcoin-qt

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18213