[cmake] deprecate BUILD_BITCOIN_X flags in favor of BUILD_X flags
Summary:
This change will make it easier to manually type `cmake` commands. The old flags are still supported for now, but a deprecation warning is printed. And if both an old and new flas are set and are not identical, the build with error.

The BUILD_BITCOIN_X flag made sense for some features that result in a  `bitcoin-x` executable being generated, but it wasn't used consistently only for this particular scenario. For instance there is no `bitcoin-chronik` or `bitcoin-zmq` executable. Also we use the same flag `BUILD_BITCOIN_WALLET` both for enabling the wallet feature in `bitcoind` and for selecting the `bitcoin-wallet`.

Test Plan:
Succeeds:
```
cmake .. -GNinja
cmake .. -GNinja -DBUILD_CHRONIK=1
```

Succeeds with a warning:
```
cmake .. -GNinja -DBUILD_BITCOIN_CHRONIK=1
cmake .. -GNinja  -DBUILD_BITCOIN_CHRONIK=ON -DBUILD_CHRONIK=ON
```

Fails:
```
cmake .. -GNinja  -DBUILD_BITCOIN_CHRONIK=1 -DBUILD_CHRONIK=0

```

check all the CI build configurations

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18231