[chronik] Unbreak the build when the old variables are used
Summary:
For builds that still use the BUILD_BITCOIN_CHRONIK build option, the new BUILD_CHRONIK value is set after the config directory is included, which causes the ENABLE_CHRONIK preprocessor to be off in the generated bitcoin-config.h. This diff defines the build value first so the header file is correct.
This doesn't prevent from building but the node bindings are not build without this change, and starting a node with -chronik=1 fails (unknown option). Note that this works just fine if the new build option is used.

Test Plan: Build with the old chronik build flag and run the functional tests.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18236