[chronik] Electrum: fix headers notifications under deep reorgs
Summary: Should a deep reorg occur, the intermediate headers might be missing because the node block index already unrolled to the new tip. This is not a problem as we want to only notify about the tip in this case. So rather than breaking the notification loop when the header is missing we can simply ignore it and we will eventually reach the tip and send it as expected.

Test Plan:
  ./test/functional/test_runner.py chronik_electrum_blockchain

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18291