Move ContextualCheckTransactionForCurrentBlock to tx_verify.h/cpp
Summary:
The other interface ContextualCheckTransaction is already here so this is where it belongs. This will make it possible to call the checks from the mempool without introducing a circular dependency.

The documentation for the function has been updated to remove an out-of-date comment. There is no change in behavior.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18363