[e.cash] Fix blog issues
Summary:
Using a variable value for the revalidate export is not allowed apparently
Needs to be static or will get build failures, so updating the three instances of it around the blog content

Also the blog images were not coming through in the build/preview environment. Need the sharp package to make it work and something is
going on with the next image optimization and the strapi images. Adding the unoptimized prop fixes it. Possibly has some negative performance impacts, but can troubleshoot further with the actual prod environment as may be a non issue on the non preview server

Test Plan: Try building on master and here. Should work here

Reviewers: #bitcoin_abc, bytesofman

Reviewed By: #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D18399