[avalanche] Properly clean up the finalized transactions from the ancestor of the finalized block
Summary:
Make sure that we remove the finalized transaction that are confirmed in a block which is an ancestor of the finalized block.
Otherwise it is possible that several blocks are mined in a row and the tip finalized, causing the transactions from the previous block to remain in the radix tree.

Depends on D18414.

Test Plan:
  ninja check check-functional

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: Mengerian, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18407