[avalanche][chronik] Bridge the isAvalancheFinalizedPreconsensus method and use it for the tx is_final status
Summary:
This method makes it possible for chronik to retrieve the finalization status from avalanche pre-consensus for transactions. This means that a tx can be shown as "is_final": true before it is included in a final block.

This is a big enough of a change but it's mostly boilerplate.

Depends on D18419.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: tobias_ruck, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D18420