Add the spent coins to the tx invalidated validation callback
Summary: This extends the callback so that chronik can retrieve the token/plugin info without being able to look up its db. This is a similar requirement to the transaction added to mempool callback where chronik needs to process the data before the tx is added to its database (in this case it's been removed already).

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, PiRK, tobias_ruck

Reviewed By: #bitcoin_abc, PiRK, tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D18438