[chronik-client] Remove rewire dependency and sorting strategy tests
Summary:
While the sorting strategy tests are nice, the functionality is not complex. Testing it appropriately, though, is complex, requiring mocked internal class methods or multiple instances with different latencies.

Since node 22 conflicts with rewire, I think the cleanest approach is to simply lose rewire and these tests. The tests are mostly confirming that "sort" is called.

We could keep the tests but not without a more complex mocking arrangement. We are already mocking so much that the test is ... just barely confirming more than its own mocks.

In this case I think removing the tests is the best approach.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18483